+++ src/test/java/org/hosh/runtime/InterpreterTest.java
@@ -24,7 +24,7 @@
 package org.hosh.runtime;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.hosh.testsupport.ExitStatusAssert.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 
@@ -36,6 +36,7 @@ import java.util.Map;
 import java.util.concurrent.CancellationException;
 
 import org.hosh.runtime.Compiler.Program;
+import org.hosh.runtime.Compiler.Resolvable;
 import org.hosh.runtime.Compiler.Statement;
 import org.hosh.spi.Channel;
 import org.hosh.spi.Command;
@@ -46,10 +47,12 @@ import org.hosh.spi.State;
 import org.hosh.spi.StateAware;
 import org.hosh.spi.TerminalAware;
 import org.hosh.spi.Values;
+import org.hosh.testsupport.WithThread;
 import org.jline.terminal.Terminal;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.Spy;
@@ -58,13 +61,19 @@ import org.mockito.junit.jupiter.MockitoExtension;
 @ExtendWith(MockitoExtension.class)
 public class InterpreterTest {
 
+	@RegisterExtension
+	public final WithThread withThread = new WithThread();
+
 	@Mock(stubOnly = true)
 	private State state;
 
 	@Mock(stubOnly = true)
 	private Terminal terminal;
 
-	@Mock(stubOnly = true)
+	@Mock
+	private Channel in;
+
+	@Mock
 	private Channel out;
 
 	@Mock
@@ -87,7 +96,7 @@ public class InterpreterTest {
 
 	private final Map<String, String> variables = new HashMap<>();
 
-	private final List<String> args = new ArrayList<>();
+	private final List<Resolvable> args = new ArrayList<>();
 
 	private Interpreter sut;
 
@@ -165,39 +174,69 @@ public class InterpreterTest {
 	}
 
 	@Test
-	public void plainArguments() {
-		args.add("file");
+	public void constantArgument() {
+		args.add(new Compiler.Constant("file"));
 		given(state.getVariables()).willReturn(variables);
 		given(command.run(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).willReturn(ExitStatus.success());
 		given(program.getStatements()).willReturn(Arrays.asList(statement));
 		given(statement.getCommand()).willReturn(command);
 		given(statement.getArguments()).willReturn(args);
 		sut.eval(program);
-		then(command).should().run(Mockito.eq(args), Mockito.any(), Mockito.any(), Mockito.any());
+		then(command).should().run(Mockito.eq(Arrays.asList("file")), Mockito.any(), Mockito.any(), Mockito.any());
 	}
 
 	@Test
 	public void presentVariable() {
-		args.add("${VARIABLE}");
+		args.add(new Compiler.Variable("VARIABLE"));
 		variables.put("VARIABLE", "1");
 		given(state.getVariables()).willReturn(variables);
 		given(command.run(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).willReturn(ExitStatus.success());
 		given(program.getStatements()).willReturn(Arrays.asList(statement));
 		given(statement.getCommand()).willReturn(command);
 		given(statement.getArguments()).willReturn(args);
-		sut.eval(program);
+		ExitStatus exitStatus = sut.eval(program);
+		assertThat(exitStatus).isSuccess();
+		then(in).shouldHaveZeroInteractions();
+		then(out).shouldHaveZeroInteractions();
+		then(err).shouldHaveZeroInteractions();
 		then(command).should().run(Mockito.eq(Arrays.asList("1")), Mockito.any(), Mockito.any(), Mockito.any());
 	}
 
 	@Test
 	public void absentVariables() {
-		args.add("${VARIABLE}");
+		args.add(new Compiler.Variable("VARIABLE"));
+		variables.put("VARIABLE", null);
 		given(state.getVariables()).willReturn(variables);
 		given(program.getStatements()).willReturn(Arrays.asList(statement));
+		given(statement.getCommand()).willReturn(command);
 		given(statement.getArguments()).willReturn(args);
-		assertThatThrownBy(() -> sut.eval(program))
-				.isInstanceOf(IllegalStateException.class)
-				.hasMessageContaining("unknown variable: VARIABLE");
+		ExitStatus exitStatus = sut.eval(program);
+		assertThat(exitStatus).isError();
+		then(in).shouldHaveZeroInteractions();
+		then(out).shouldHaveZeroInteractions();
+		then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("cannot resolve variable: VARIABLE")));
+	}
+
+	@Test
+	public void setThreadNameWithArgs() {
+		given(command.describe()).willReturn("java");
+		given(command.run(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).willReturn(ExitStatus.success());
+		given(statement.getCommand()).willReturn(command);
+		given(statement.getArguments()).willReturn(Arrays.asList(new Compiler.Constant("-jar"), new Compiler.Constant("hosh.jar")));
+		sut.run(statement, in, out, err);
+		assertThat(Thread.currentThread().getName()).isEqualTo("command='java -jar hosh.jar'");
+		then(err).shouldHaveZeroInteractions(); // checking no assertion failures happened
+	}
+
+	@Test
+	public void setThreadNameWithoutArgs() {
+		given(command.describe()).willReturn("java");
+		given(command.run(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).willReturn(ExitStatus.success());
+		given(statement.getCommand()).willReturn(command);
+		given(statement.getArguments()).willReturn(Arrays.asList());
+		sut.run(statement, in, out, err);
+		assertThat(Thread.currentThread().getName()).isEqualTo("command='java'");
+		then(err).shouldHaveNoMoreInteractions(); // checking no assertion failures happened
 	}
 
 	public interface StateAwareCommand extends Command, StateAware {